+++ src/main/java/com/example/linkcargo/domain/chat/Entity/Membership.java
@@ -1,6 +1,5 @@
 package com.example.linkcargo.domain.chat.Entity;
 
-import com.example.linkcargo.domain.chat.ChatRoom;
 import com.example.linkcargo.domain.user.User;
 import com.example.linkcargo.global.entity.JpaBaseEntity;
 import jakarta.persistence.Entity;
