+++ src/main/java/com/example/linkcargo/domain/chat/dto/ChatContentResponse.java
@@ -1,0 +1,7 @@
+package com.example.linkcargo.domain.chat.dto;
+
+public record ChatContentResponse(
+    Long chatRoomId,
+    String content
+) {
+}
