+++ src/main/java/com/example/linkcargo/domain/chat/dto/ChatEnterResponse.java
@@ -1,0 +1,6 @@
+package com.example.linkcargo.domain.chat.dto;
+
+public record ChatEnterResponse(
+    Long chatRoomId
+) {
+}
