+++ src/main/java/com/example/linkcargo/domain/chat/repository/ChatRoomRepository.java
@@ -1,6 +1,6 @@
 package com.example.linkcargo.domain.chat.repository;
 
-import com.example.linkcargo.domain.chat.ChatRoom;
+import com.example.linkcargo.domain.chat.Entity.ChatRoom;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
